+++ lib/levanter/src/levanter/grug/_moe/__init__.py
@@ -1,0 +1,4 @@
+# Copyright The Levanter Authors
+# SPDX-License-Identifier: Apache-2.0
+
+"""Private Grug MoE backend implementations."""
